requirements.txt
torch>=2.0
torchvision
numpy
scipy
pandas
matplotlib
opencv-python
PyYAML
tqdm
scikit-learn
